보상 트랜잭션 포인트 사용 취소 구현

+++ user-service/src/main/kotlin/org/ktor_lecture/userservice/application/service/PointService.kt
@@ -110,7 +110,7 @@ class PointService (
      */
     @Transactional
     override fun cancel(command: PointCancelCommand) {
-        val user = userReadRepository.findById(command.userId.toLong()).orElseThrow { throw ConcertException(ErrorCode.USER_NOT_FOUND) }
+        val user = userReadRepository.findById(command.userId).orElseThrow { throw ConcertException(ErrorCode.USER_NOT_FOUND) }
         val pointHistory = pointHistoryRepository.findById(command.pointHistoryId).orElseThrow { throw ConcertException(ErrorCode.POINT_HISTORY_NOT_FOUND) }
